pyproject.toml
[project]
name = "titiler.stacapi"
description = "Connect titiler to STAC APIs."
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
authors = [
    {name = "Vincent Sarago", email = "vincent@developmentseed.com"},
]
keywords = [
    "TiTiler",
    "STAC API",
]
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]
dependencies = [
    "orjson",
    "titiler.core>=0.17.0,<0.19",
    "titiler.mosaic>=0.17.0,<0.19",
    "rio_tiler@git+https://github.com/cogeotiff/rio-tiler#egg=88d7ccc8bad697471e1d61ab982ecc9501633a63",
    "pystac-client",
    "pydantic>=2.4,<3.0",
    "pydantic-settings~=2.0",
    "planetary_computer~=1.0.0",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pre-commit",
]
test = [
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "httpx",
]

[project.urls]
Homepage = 'https://developmentseed.org/titiler-stacapi/'
Documentation = "https://developmentseed.org/titiler-stacapi/"
Issues = "https://github.com/developmentseed/titiler-stacapi/issues"
Source = "https://github.com/developmentseed/titiler-stacapi"
Changelog = "https://developmentseed.org/titiler-stacapi/release-notes/"

[build-system]
requires = ["pdm-pep517"]
build-backend = "pdm.pep517.api"

[tool.pdm.version]
source = "file"
path = "titiler/stacapi/__init__.py"

[tool.pdm.build]
includes = ["titiler/stacapi"]
excludes = ["tests/", "**/.mypy_cache", "**/.DS_Store"]

[tool.coverage.run]
branch = true
parallel = true

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.isort]
profile = "black"
known_first_party = ["titiler"]
known_third_party = [
    "rasterio",
    "morecantile",
    "rio_tiler",
    "cogeo_mosaic",
    "geojson_pydantic",
]
default_section = "THIRDPARTY"

[tool.ruff]
select = [
    "D1",  # pydocstyle errors
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # flake8
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "B905",  # ignore zip() without an explicit strict= parameter, only support with python >3.10
]

[tool.mypy]
no_implicit_optional = true
strict_optional = true
namespace_packages = true
explicit_package_bases = true
